#292 - crux: reorganize image stuff
