--- src/crux/allocators/mod.rs
@@ -1,6 +1,0 @@
-//  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
-//  SPDX-License-Identifier: MIT
-
-//! env heap allocators
-pub mod bump_allocator;
-pub mod stack_allocator;
