+++ src/crux/images/allocator.rs
@@ -3,19 +3,20 @@
 
 //! env allocators
 //!    Env
-use futures::executor::block_on;
 use {
     crate::{
-        allocators::{bump_allocator::BumpAllocator, stack_allocator::StackAllocator},
         core::{
             env::Env,
             heap::Heap,
             types::{Tag, Type},
         },
+        images::{bump_allocator::BumpAllocator, stack_allocator::StackAllocator},
     },
     modular_bitfield::specifiers::{B11, B4},
     std::fmt,
 };
+
+use futures::executor::block_on;
 
 #[bitfield]
 #[repr(align(8))]
