+++ src/crux/images/mod.rs
@@ -1,2 +1,4 @@
+pub mod allocator;
 pub mod bump_allocator;
 pub mod image;
+pub mod stack_allocator;
